refactor: to put logic in the providers folder
