+++ crates/lsp/src/providers/uri.rs
@@ -1,0 +1,2 @@
+// This module is reserved for cross-platform URI handling utilities.
+// Functions can be added here as needed.
